Help entries

These entries can be viewed at the Julia REPL Julia using

julia> ? 
help?> function_name

CellListMap.AuxThreaded — Type

struct AuxThreaded{N, T}

struct AuxThreaded{N, T}

Auxiliary structure to carry threaded lists and ranges of particles to be considered by each thread on parallel construction.

CellListMap.AuxThreaded — Method

AuxThreaded(cl::CellListPair{N,T}) where {N,T}

Constructor for the AuxThreaded type for lists of disjoint particle sets, to be passed to UpdateCellList! for in-place update of cell lists.

Example

julia> box = Box([250,250,250],10);

julia> x = [ 250*rand(3) for i in 1:50_000 ];

julia> y = [ 250*rand(3) for i in 1:10_000 ];

julia> cl = CellList(x,y,box);

julia> aux = CellListMap.AuxThreaded(cl)
CellListMap.AuxThreaded{3, Float64}
 Auxiliary arrays for nthreads = 8

julia> cl = UpdateCellList!(x,box,cl,aux)
CellList{3, Float64}
  100000 real particles.
  31190 cells with real particles.
  1134378 particles in computing box, including images.

CellListMap.AuxThreaded — Method

AuxThreaded(cl::CellList{N,T}) where {N,T}

Constructor for the AuxThreaded type, to be passed to UpdateCellList! for in-place update of cell lists.

Example

julia> box = Box([250,250,250],10);

julia> x = [ 250*rand(3) for _ in 1:100_000 ];

julia> cl = CellList(x,box);

julia> aux = CellListMap.AuxThreaded(cl)
CellListMap.AuxThreaded{3, Float64}
 Auxiliary arrays for nthreads = 8

julia> cl = UpdateCellList!(x,box,cl,aux)
CellList{3, Float64}
  100000 real particles.
  31190 cells with real particles.
  1134378 particles in computing box, including images.

CellListMap.Box — Type

struct Box{UnitCellType, N, T, M}

unit_cell::CellListMap.UnitCell

lcell::Int64

nc::StaticArrays.SVector{N, Int64} where N

cutoff::Any

cutoff_sq::Any

ranges::StaticArrays.SVector{N, UnitRange{Int64}} where N

cell_size::StaticArrays.SVector{N, T} where {N, T}

unit_cell_max::StaticArrays.SVector{N, T} where {N, T}

Structure that contains some data required to compute the linked cells. To be initialized with the box size and cutoff.

Examples

julia> using CellListMap

julia> sides = [250,250,250];

julia> cutoff = 10;

julia> box = Box(sides,cutoff)
Box{OrthorhombicCell, 3, Float64, 9}
  unit cell matrix: [250.0 0.0 0.0; 0.0 250.0 0.0; 0.0 0.0 250.0]
  cutoff: 10.0
  number of computing cells on each dimension: [27, 27, 27]
  computing cell sizes: [10.0, 10.0, 10.0] (lcell: 1)
  Total number of cells: 19683

julia> box = Box([ 10  0  0 
                    0 10  5
                    0  0 10 ], 1)
Box{TriclinicCell, 3, Float64, 9}
unit cell matrix: [10.0 0.0 0.0; 0.0 10.0 5.0; 0.0 0.0 10.0]
cutoff: 1.0
number of computing cells on each dimension: [12, 17, 12]
computing cell sizes: [1.0, 1.0, 1.0] (lcell: 1)
Total number of cells: 2448

CellListMap.Box — Method

Box(
    limits::Limits,
    cutoff;
    T::DataType=Float64,
    lcell::Int=1
)

This constructor receives the output of limits(x) or limits(x,y) where x and y are the coordinates of the particles involved, and constructs a Box with size larger than the maximum coordinates ranges of all particles plus the cutoff. This is used to emulate pairwise interactions in non-periodic boxes. The output box is always an Orthorhombic cell.

Examples

julia> x = [ [100,100,100] .* rand(3) for i in 1:100_000 ];

julia> box = Box(limits(x),10)
Box{OrthorhombicCell, 3, Float64, 9}
  unit cell matrix: [109.99633932875878 0.0 0.0; 0.0 109.99780283179763 0.0; 0.0 0.0 109.99587254766517]
  cutoff: 10.0
  number of computing cells on each dimension: [12, 12, 12]
  computing cell sizes: [10.999633932875877, 10.999780283179764, 10.999587254766517] (lcell: 1)
  Total number of cells: 1728

julia> y = [ [150,150,50] .* rand(3) for i in 1:100_000 ];

julia> box = Box(limits(x,y),10)
Box{OrthorhombicCell, 3, Float64, 9}
  unit cell matrix: [159.99787690924168 0.0 0.0; 0.0 159.98878289444897 0.0; 0.0 0.0 109.99587254766517]
  cutoff: 10.0
  number of computing cells on each dimension: [18, 17, 12]
  computing cell sizes: [10.666525127282778, 10.665918859629931, 10.999587254766517] (lcell: 1)
  Total number of cells: 3672

CellListMap.Box — Method

Box(
  unit_cell_matrix::AbstractMatrix, 
  cutoff, 
  T::DataType, 
  lcell::Int=1,
  UnitCellType=TriclinicCell
)

Construct box structure given the cell matrix of lattice vectors. This constructor will always return a TriclinicCell box type, unless the UnitCellType parameter is set manually to OrthorhombicCell

Example

julia> unit_cell = [ 100   50    0 
                       0  120    0
                       0    0  130 ];

julia> box = Box(unit_cell,10)
Box{TriclinicCell, 3, Float64, 9}
  unit cell matrix: [100.0 50.0 0.0; 0.0 120.0 0.0; 0.0 0.0 130.0]
  cutoff: 10.0
  number of computing cells on each dimension: [17, 14, 15]
  computing cell sizes: [10.0, 10.0, 10.0] (lcell: 1)
  Total number of cells: 3570

CellListMap.Box — Method

Box(
  sides::AbstractVector, 
  cutoff, 
  T::DataType, 
  lcell::Int=1,
  UnitCellType=OrthorhombicCell
)

For orthorhombic unit cells, Box can be initialized with a vector of the length of each side.

Example

julia> box = Box([120,150,100],10)
Box{OrthorhombicCell, 3, Float64, 9}
  unit cell matrix: [120.0 0.0 0.0; 0.0 150.0 0.0; 0.0 0.0 100.0]
  cutoff: 10.0
  number of computing cells on each dimension: [14, 17, 12]
  computing cell sizes: [10.0, 10.0, 10.0] (lcell: 1)
  Total number of cells: 2856

CellListMap.Cell — Type

struct Cell{N, T}

linear_index::Int64

cartesian_index::CartesianIndex

center::StaticArrays.SVector{N, T} where {N, T}

contains_real::Bool

n_particles::Int64

particles::Array{CellListMap.ParticleWithIndex{N, T}, 1} where {N, T}

This structure contains the cell linear index and the information about if this cell is in the border of the box (such that its neighbouring cells need to be wrapped)

CellListMap.CellList — Type

struct CellList{N, T}

n_real_particles::Int64

Number of real particles.

number_of_cells::Int64

Number of cells.

n_particles::Int64

mutable number of particles in the computing box.

n_cells_with_real_particles::Int64

mutable number of cells with real particles.

n_cells_with_particles::Int64

mutable number of cells with particles, real or images.

cell_indices::Vector{Int64}

Auxiliary array that contains the indexes in list of the cells with particles, real or images.

cell_indices_real::Vector{Int64}

Auxiliary array that contains the indexes in the cells with real particles.

cells::Array{CellListMap.Cell{N, T}, 1} where {N, T}

Vector containing cell lists of cells with particles.

projected_particles::Array{Array{CellListMap.ProjectedParticle{N, T}, 1}, 1} where {N, T}

Auxiliar array to store projected particles.

Structure that contains the cell lists information.

CellListMap.CellList — Method

function CellList(
    x::AbstractMatrix,
    y::AbstractMatrix,
    box::Box{UnitCellType,N,T};
    parallel::Bool=true,
    autoswap=true
) where {UnitCellType,N,T} 

Reinterprets the matrices x and y as vectors of static vectors and calls the equivalent function with the reinterprted input. The first dimension of the matrices must be the dimension of the points (2 or 3).

CellListMap.CellList — Method

function CellList(
    x::AbstractMatrix,
    box::Box{UnitCellType,N,T};
    parallel::Bool=true
) where {UnitCellType,N,T} 

Reinterprets the matrix x as vectors of static vectors and calls the equivalent function with the reinterprted input. The first dimension of the matrix must be the dimension of the points (2 or 3).

CellListMap.CellList — Method

CellList(
    x::AbstractVector{<:AbstractVector},
    y::AbstractVector{<:AbstractVector},
    box::Box{UnitCellType,N,T};
    parallel::Bool=true,
    autoswap::Bool=true
) where {UnitCellType,N,T} 

Function that will initialize a CellListPair structure from scracth, given two vectors of particle coordinates and a Box, which contain the size of the system, cutoff, etc. By default, the cell list will be constructed for smallest vector, but this is not always the optimal choice. Using autoswap=false the cell list is constructed for the second (y)

Example

julia> box = Box([250,250,250],10);

julia> x = [ 250*rand(SVector{3,Float64}) for i in 1:1000 ];

julia> y = [ 250*rand(SVector{3,Float64}) for i in 1:10000 ];

julia> cl = CellList(x,y,box)
CellListMap.CellListPair{Vector{SVector{3, Float64}}, 3, Float64}
   10000 particles in the reference vector.
   961 cells with real particles of target vector.

julia> cl = CellList(x,y,box,autoswap=false)
CellListMap.CellListPair{Vector{SVector{3, Float64}}, 3, Float64}
   1000 particles in the reference vector.
   7389 cells with real particles of target vector.

CellListMap.CellList — Method

CellList(
    x::AbstractVector{AbstractVector},
    box::Box{UnitCellType,N,T};
    parallel::Bool=true
) where {UnitCellType,N,T} 

Function that will initialize a CellList structure from scracth, given a vector or particle coordinates (a vector of vectors, typically of static vectors) and a Box, which contain the size ofthe system, cutoff, etc.

Example

julia> box = Box([250,250,250],10);

julia> x = [ 250*rand(SVector{3,Float64}) for i in 1:100000 ];

julia> cl = CellList(x,box)
CellList{3, Float64}
  100000 real particles.
  15600 cells with real particles.
  126276 particles in computing box, including images.

CellListMap.CellListPair — Type

struct CellListPair{V, N, T}

ref::Any

target::CellList

swap::Bool

Structure that will cointain the cell lists of two independent sets of particles for cross-computation of interactions

CellListMap.Limits — Type

struct Limits{T<:(AbstractVector{T} where T)}

limits::AbstractVector{T} where T

Structure that contains the maximum lengths on each direction, to dispatch on the construction of boxes without periodic boundary conditions.

CellListMap.ParticleWithIndex — Type

struct ParticleWithIndex{N, T}

index::Int32

real::Bool

coordinates::StaticArrays.SVector{N, T} where {N, T}

Copies particle coordinates and associated index, to build contiguous particle lists in memory when building the cell lists. This strategy duplicates the particle coordinates data, but is probably worth the effort. The index is a 32bit integer such that the complete struct has 32bytes.

CellListMap.ProjectedParticle — Type

struct ProjectedParticle{N, T}

index::Int32

xproj::Float32

coordinates::StaticArrays.SVector{N, T} where {N, T}

Auxiliary structure to contain projected particles. Types of scalars are chosen such that with a SVector{3,Float64} the complete struct has 32bytes.

CellListMap.UpdateCellList! — Method

function UpdateCellList!(
    x::AbstractMatrix,
    box::Box,
    cl::CellList{N,T},
    aux::AuxThreaded{N,T};
    parallel::Bool=true
) where {N,T}

Reinterprets the matrix x as vectors of static vectors and calls the equivalent function with the reinterprted input. The first dimension of the matrix must be the dimension of the points (2 or 3).

CellListMap.UpdateCellList! — Method

function UpdateCellList!(
    x::AbstractMatrix,
    box::Box,
    cl::CellList{N,T};
    parallel::Bool=true
) where {N,T}

Reinterprets the matrix x as vectors of static vectors and calls the equivalent function with the reinterprted input. The first dimension of the matrix must be the dimension of the points (2 or 3).

CellListMap.UpdateCellList! — Method

function UpdateCellList!(
    x::AbstractVector{<:AbstractVector},
    box::Box,
    cl::CellList{N,T},
    aux::AuxThreaded{N,T};
    parallel::Bool=true
) where {N,T}

Function that updates the cell list cl new coordinates x and possibly a new box box, and receives a preallocated aux structure of auxiliary vectors for threaded cell list construction. Given a preallocated aux vector, allocations in this function should be minimal, only associated with the spawning threads, or to expansion of the cell lists if the number of cells or number of particles increased.

Example

julia> box = Box([250,250,250],10);

julia> x = [ 250*rand(SVector{3,Float64}) for i in 1:100000 ];

julia> cl = CellList(x,box);

julia> aux = CellListMap.AuxThreaded(cl)
CellListMap.AuxThreaded{3, Float64}
 Auxiliary arrays for nthreads = 8

julia> x = [ 250*rand(SVector{3,Float64}) for i in 1:100000 ];

julia> UpdateCellList!(x,box,cl,aux)
CellList{3, Float64}
  100000 real particles.
  15599 cells with real particles.
  125699 particles in computing box, including images.

To illustrate the expected ammount of allocations, which are a consequence of thread spawning only:

julia> using BenchmarkTools

julia> @btime UpdateCellList!($x,$box,$cl,$aux)
  16.384 ms (41 allocations: 3.88 KiB)
CellList{3, Float64}
  100000 real particles.
  15599 cells with real particles.
  125699 particles in computing box, including images.

julia> @btime UpdateCellList!($x,$box,$cl,$aux,parallel=false)
  20.882 ms (0 allocations: 0 bytes)
CellList{3, Float64}
  100000 real particles.
  15603 cells with real particles.
  125896 particles in computing box, including images.

CellListMap.UpdateCellList! — Method

UpdateCellList!(
    x::AbstractVector{<:AbstractVector},
    box::Box,
    cl:CellList{N,T},
    parallel=true
) where {N,T}

Function that will update a previously allocated CellList structure, given new updated particle positions. This function will allocate new threaded auxiliary arrays in parallel calculations. To preallocate these auxiliary arrays, use the UpdateCellList!(x,box,cl,aux) method instead.

Example

julia> box = Box([250,250,250],10);

julia> x = [ 250*rand(SVector{3,Float64}) for i in 1:1000 ];

julia> cl = CellList(x,box);

julia> box = Box([260,260,260],10);

julia> x = [ 260*rand(SVector{3,Float64}) for i in 1:1000 ];

julia> cl = UpdateCellList!(x,box,cl); # update lists

CellListMap.UpdateCellList! — Method

function UpdateCellList!(
    x::AbstractMatrix,
    y::AbstractMatrix,
    box::Box{UnitCellType,N,T},
    cl_pair::CellListPair,
    aux::AuxThreaded{N,T};
    parallel::Bool=true
) where {UnitCellType,N,T}

Reinterprets the matrices x and y as vectors of static vectors and calls the equivalent function with the reinterprted input. The first dimension of the matrices must be the dimension of the points (2 or 3).

CellListMap.UpdateCellList! — Method

function UpdateCellList!(
    x::AbstractMatrix,
    y::AbstractMatrix,
    box::Box{UnitCellType,N,T},
    cl_pair::CellListPair;
    parallel::Bool=true
) where {UnitCellType,N,T}

Reinterprets the matrices x and y as vectors of static vectors and calls the equivalent function with the reinterprted input. The first dimension of the matrices must be the dimension of the points (2 or 3).

CellListMap.UpdateCellList! — Method

function UpdateCellList!(
    x::AbstractVector{<:AbstractVector},
    y::AbstractVector{<:AbstractVector},
    box::Box{UnitCellType,N,T},
    cl_pair::CellListPair,
    aux::AuxThreaded{N,T};
    parallel::Bool=true
) where {UnitCellType,N,T}

This function will update the cl_pair structure that contains the cell lists for disjoint sets of particles. It receives the preallocated aux structure to avoid reallocating auxiliary arrays necessary for the threaded construct of the lists.

Example

julia> box = Box([250,250,250],10);

julia> x = [ 250*rand(3) for i in 1:50_000 ];

julia> y = [ 250*rand(3) for i in 1:10_000 ];

julia> cl = CellList(x,y,box)
CellListMap.CellListPair{Vector{SVector{3, Float64}}, 3, Float64}
   50000 particles in the reference vector.
   7381 cells with real particles of target vector.

julia> aux = CellListMap.AuxThreaded(cl)
CellListMap.AuxThreaded{3, Float64}
 Auxiliary arrays for nthreads = 8

julia> x = [ 250*rand(3) for i in 1:50_000 ];

julia> y = [ 250*rand(3) for i in 1:10_000 ];

julia> cl = UpdateCellList!(x,y,box,cl,aux)
CellList{3, Float64}
  10000 real particles.
  7358 cells with real particles.
  12591 particles in computing box, including images.

To illustrate the expected ammount of allocations, which are a consequence of thread spawning only:

julia> using BenchmarkTools

julia> @btime UpdateCellList!($x,$y,$box,$cl,$aux)
  715.661 μs (41 allocations: 3.88 KiB)
CellListMap.CellListPair{Vector{SVector{3, Float64}}, 3, Float64}
   50000 particles in the reference vector.
   7414 cells with real particles of target vector.
   
julia> @btime UpdateCellList!($x,$y,$box,$cl,$aux,parallel=false)
   13.042 ms (0 allocations: 0 bytes)
 CellListMap.CellListPair{Vector{SVector{3, Float64}}, 3, Float64}
    50000 particles in the reference vector.
    15031 cells with real particles of target vector.
 

CellListMap.UpdateCellList! — Method

UpdateCellList!(
  x::AbstractVector{<:AbstractVector},
  y::AbstractVector{<:AbstractVector},
  box::Box{UnitCellType,N,T},
  cl:CellListPair,
  parallel=true
) where {UnitCellType,N,T}

Function that will update a previously allocated CellListPair structure, given new updated particle positions, for example. This method will allocate new aux threaded auxiliary arrays. For a non-allocating version, see the UpdateCellList!(x,y,box,cl,aux) method.

julia> box = Box([250,250,250],10);

julia> x = [ 250*rand(SVector{3,Float64}) for i in 1:1000 ];

julia> y = [ 250*rand(SVector{3,Float64}) for i in 1:10000 ];

julia> cl = CellList(x,y,box);

julia> cl = UpdateCellList!(x,y,box,cl); # update lists

CellListMap.add_particle_to_celllist! — Method

add_particle_to_celllist!(
    ip,
    x::SVector{N,T},
    box,
    cl::CellList{N,T};
    real_particle::Bool=true
) where {N,T}

Adds one particle to the cell lists, updating all necessary arrays.

CellListMap.add_particles! — Method

add_particles!(x,box,ishift,cl::CellList{N,T}) where {N,T}

Add all particles in vector x to the cell list cl. ishift is the shift in particle index, meaning that particle i of vector x corresponds to the particle with original index i+ishift. The shift is used to construct cell lists from fractions of the original set of particles in parallel list construction.

CellListMap.append_particles! — Method

append_particles!(cell1::Cell,cell2::Cell)

Add the particles of cell2 to cell1, updating the cell data and, if necessary, resizing (increasing) the particles array of cell1

CellListMap.cell_cartesian_indices — Method

cell_cartesian_indices(nc::SVector{N,Int}, i1D) where {N}

Given the linear index of the cell in the cell list, returns the cartesian indices of the cell (for arbitrary dimension N).

CellListMap.cell_center — Method

cell_center(c::CartesianIndex{N},box::Box{UnitCellType,N,T}) where {UnitCellType,N,T}

Computes the geometric center of a computing cell, to be used in the projection of points. Returns a SVector{N,T}

CellListMap.cell_linear_index — Method

cell_linear_index(nc::SVector{N,Int}, indices) where N

Returns the index of the cell, in the 1D representation, from its cartesian coordinates.

CellListMap.check_unit_cell — Method

check_unit_cell(box::Box)

Checks if the unit cell satisfies the conditions for using the minimum-image convention.

CellListMap.copydata! — Method

copydata!(cell1::Cell,cell2::Cell)

Copies the data from cell2 to cell1, meaning that particles are copied element-wise from cell2 to cell1, with the particles array of cell1 being resized (increased) if necessary.

CellListMap.dot — Method

dot(x::AbstractVector{T1},y::AbstractVector{T2}) where {T1,T2}

LinearAlgebra.dot is slower than this for standard arrays (likely more accurate, but that is not relevant here).

CellListMap.draw_computing_cell — Method

draw_computing_cell(x,box::Box{UnitCellType,2}) where UnitCellType

draw_computing_cell(cl::CellList,box::Box{UnitCellType,2},x) where UnitCellType

This function creates a plot of the computing cell, in two dimensions.

CellListMap.draw_computing_cell — Method

draw_computing_cell(x,box::Box{UnitCellType,3}) where UnitCellType

This function creates a plot of the computing cell, in three dimensions.

CellListMap.fix_upper_boundary — Method

fix_upper_boundary(x::T,side) where T

Move x to x -side if x == side, because we use the convention that the boundary belongs to the next cell.

CellListMap.init_aux_threaded! — Method

init_aux_threaded!(aux::AuxThreaded,cl::CellList)

Given an AuxThreaded object initialized with zero-length arrays, push ntrheads copies of cl into aux.lists and resize aux.idxs to the number of threads.

CellListMap.limits — Method

limits(x::AbstractVector{<:AbstractVector})

Returns the lengths of a orthorhombic box that encompasses all the particles defined in x, to be used to set a box without effective periodic boundary conditions.

CellListMap.limits — Method

limits(x::T,y::T) where T<:AbstractVector{<:AbstractVector})

Returns the lengths of a orthorhombic box that encompasses all the particles defined in x and y, to used to set a box without effective periodic boundary conditions.

CellListMap.map_naive! — Method

map_naive!(f,output,x,box)

Function that uses the naive pairwise mapping algorithm, for testing.

CellListMap.map_pairwise! — Method

map_pairwise!(f::Function,output,box::Box,cl::CellListPair)

The same but to evaluate some function between pairs of the particles of the vectors.

CellListMap.map_pairwise! — Method

map_pairwise!(
  f::Function,
  output,
  box::Box,
  cl::CellList
  ;parallel::Bool=true,
  show_progress::Bool=false
)

This function will run over every pair of particles which are closer than box.cutoff and compute the Euclidean distance between the particles, considering the periodic boundary conditions given in the Box structure. If the distance is smaller than the cutoff, a function f of the coordinates of the two particles will be computed.

The function f receives six arguments as input:

f(x,y,i,j,d2,output)

Which are the coordinates of one particle, the coordinates of the second particle, the index of the first particle, the index of the second particle, the squared distance between them, and the output variable. It has also to return the same output variable. Thus, f may or not mutate output, but in either case it must return it. With that, it is possible to compute an average property of the distance of the particles or, for example, build a histogram. The squared distance d2 is computed internally for comparison with the cutoff, and is passed to the f because many times it is used for the desired computation.

Example

Computing the mean absolute difference in x position between random particles, remembering the number of pairs of n particles is n(n-1)/2. The function does not use the indices or the distance, such that we remove them from the parameters by using a closure.

julia> n = 100_000;

julia> box = Box([250,250,250],10);

julia> x = [ SVector{3,Float64}(sides .* rand(3)) for i in 1:n ];

julia> cl = CellList(x,box);

julia> f(x,y,sum_dx) = sum_dx + abs(x[1] - y[1])

julia> normalization = N / (N*(N-1)/2) # (number of particles) / (number of pairs)

julia> avg_dx = normalization * map_parwise!((x,y,i,j,d2,sum_dx) -> f(x,y,sum_dx), 0.0, box, cl)

CellListMap.merge_cell_lists! — Method

merge_cell_lists!(cl::CellList,aux::CellList)

Merges an auxiliary aux cell list to cl, and returns the modified cl. Used to merge cell lists computed in parallel threads.

CellListMap.neighbour_cells — Method

neighbour_cells(box::Box{UnitCellType,N}) where N

Function that returns the iterator of the cartesian indices of all neighbouring cells of a cell if the cells have sides of box.cell_size. N can be 2 or 3, for two- or three-dimensional systems.

CellListMap.neighbour_cells_forward — Method

neighbour_cells_forward(box::Box{UnitCellType,N}) where UnitCellType 

Function that returns the iterator of the cartesian indices of the cells that must be evaluated (forward, i. e. to avoid repeated interactions) if the cells have sides of length box.cell_size. N can be 2 or 3, for two- or three-dimensional systems.

CellListMap.neighbourlist — Method

neighbourlist(x,y,r;parallel=true)

Computes the list of pairs of particles of x which are closer than r to the particles of y.

Example

julia> x = [ rand(3) for i in 1:10_000 ];

julia> y = [ rand(3) for i in 1:1_000 ];

julia> CellListMap.neighbourlist(x,y,0.05)
5006-element Vector{Tuple{Int64, Int64, Float64}}:
 (1, 269, 0.04770884036497686)
 (25, 892, 0.03850515231540869)
 ⋮
 (9952, 749, 0.048875643578313456)
 (9984, 620, 0.04101242499363183)

CellListMap.neighbourlist — Method

neighbourlist(x,r;parallel=true)

Computes the list of pairs of particles in x which are closer to each other than r.

Example

julia> x = [ rand(3) for i in 1:10_000 ];

julia> CellListMap.neighbourlist(x,0.05)
24848-element Vector{Tuple{Int64, Int64, Float64}}:
 (1, 1055, 0.022977369806392412)
 (1, 5086, 0.026650609138167428)
 ⋮
 (9989, 3379, 0.0467653507446483)
 (9989, 5935, 0.02432728985151653)

CellListMap.neighbourlist — Method

neighbourlist(box, cl; parallel=true)

Compute the neighbour list of a single set or set pairs of particles. Returns a vector of tuples with all indices of the particles that are within box.cutoff, and the distances.

Example

julia> x = [ rand(3) for i in 1:1000 ];

julia> box = Box([1,1,1],0.02) 
Box{OrthorhombicCell, 3, Float64, 9}
  unit cell matrix: [1.0 0.0 0.0; 0.0 1.0 0.0; 0.0 0.0 1.0]
  cutoff: 0.02
  number of computing cells on each dimension: [52, 52, 52]
  computing cell sizes: [0.02, 0.02, 0.02] (lcell: 1)
  Total number of cells: 140608

julia> cl = CellList(x,box) # single set
CellList{3, Float64}
  999 cells with real particles.
  1126 particles in computing box, including images.

julia> CellListMap.neighbourlist(box,cl,parallel=false)
15-element Vector{Tuple{Int64, Int64, Float64}}:
 (187, 511, 0.010346860078531755)
 (203, 708, 0.010777737363239403)
 (296, 579, 0.018124283912224655)
 ⋮
 (584, 4, 0.016935844769524398)
 (725, 749, 0.019971874892397875)
 (773, 119, 0.01835233336121765)
 (927, 8, 0.011234110402648743)

To obtain the neighbour list (within the cutoff) between two sets of particles, initialize the cell lists with the two sets:

julia> x = [ rand(3) for i in 1:1000 ];

julia> y = [ rand(3) for i in 1:1000 ];

julia> box = Box([1,1,1],0.02);

julia> cl = CellList(x,y,box)

julia> cl = CellList(x,y,box)
CellListMap.CellListPair{Vector{SVector{3, Float64}}, 3, Float64}
   1000 particles in the reference vector.
   997 cells with real particles of target vector.

julia> CellListMap.neighbourlist(box,cl)
35-element Vector{Tuple{Int64, Int64, Float64}}:
 (409, 982, 0.01634641594779082)
 (521, 422, 0.00919026348035512)
 (625, 731, 0.012986301890746663)
 ⋮
 (647, 730, 0.01565763971458105)
 (296, 668, 0.016556686306217868)
 (992, 589, 0.018392993428289553)

CellListMap.norm — Method

norm(v::AbstractVector{T}) where T

norm_sqr from LinearAlgebra is not documented and is slower than this for standard arrays. Thus we define our own norm(x) = norm_sqr(x).

CellListMap.norm_sqr — Method

norm_sqr(v::AbstractVector{T}) where T

norm_sqr from LinearAlgebra is not documented and is slower than this for standard arrays.

CellListMap.out_of_bounding_box — Method

out_of_bounding_box(x::SVector{N},box::Box) where N

Function that evaluates if a particle is outside the computing bounding box, defined by the maximum and minimum unit cell coordinates.

CellListMap.particle_cell — Method

particle_cell(x::SVector{N,T}, box::Box) where {N,T}

Returns the coordinates of the computing cell to which a particle belongs, given its coordinates and the cell_size vector. The computing box is always Orthorhombic, and the first computing box with positive coordinates has indexes Box.lcell + 1.

CellListMap.particles_per_cell — Method

particles_per_cell(cl::CellList,box::Box)

Returns the average number of particles per computing cell.

CellListMap.partition! — Method

partition!(x::AbstractVector,by)

Function that reorders x vector by putting in the first positions the elements with values satisfying by(el). Returns the number of elements that satisfy the condition.

CellListMap.project_particles! — Method

project_particles!(projected_particles,cellⱼ,cellᵢ,Δc,Δc_norm,box)

Projects all particles of the cell cellⱼ into unnitary vector Δc with direction connecting the centers of cellⱼ and cellᵢ. Modifies projected_particles, and returns a view of `projected particles, where only the particles for which the projection on the direction of the cell centers still allows the particle to be within the cutoff distance of any point of the other cell.

CellListMap.ranges_of_replicas — Method

ranges_of_replicas(cell_size, lcell, nc, unit_cell_matrix::SMatrix{3,3,T}) where T

Function that sets which is the range of periodic images necessary to fill the computing box, in 2D.

CellListMap.ranges_of_replicas — Method

ranges_of_replicas(cell_size, lcell, nc, unit_cell_matrix::SMatrix{3,3,T}) where T

Function that sets which is the range of periodic images necessary to fill the computing box, in 3D.

CellListMap.reduce — Method

reduce(output::Number, output_threaded::Vector{<:Number})

Functions to reduce the output of common options (vectors of numbers and vectors of vectors). This function can be overloaded by custom reduction methods. It always must both receive the output variable as a parameter, and return it at the end.

CellListMap.replicate_particle! — Method

replicate_particle!(ip,p::SVector{N},box,cl) where N

Replicates the particle as many times as necessary to fill the computing box.

CellListMap.reset! — Method

reset!(cl::CellList{N,T},box) where{N,T}

Resets a cell list, by setting everything to zero, but retaining the allocated particles and projected_particles vectors.

CellListMap.set_nt — Method

set_nt(cl) = max(1,min(cl.n_real_particles÷500,nthreads()))

Don't use all threads to build cell lists if the number of particles per thread is smaller than 500.

CellListMap.strip_value — Method

strip_value(x)

This function by default does nothing, but it can be overloaded to remove decorations from types of variables that contain more information that the coordinate itself. For example, it can be used to deal with Unitful coordinates internaly, without having to copy the coordinates prior to passing it to CellList. For example:

julia> box = Box([100,100,100],10);

julia> using Unitful

julia> x = [ sides .* rand(SVector{3,Float64})u"nm" for i in 1:5 ]
5-element Vector{SVector{3, Quantity{Float64, 𝐋, Unitful.FreeUnits{(nm,), 𝐋, nothing}}}}:
 [204.13761649857798 nm, 221.72338544218846 nm, 68.39256684613026 nm]
 [240.3490051044633 nm, 39.133140818690904 nm, 93.46671201972323 nm]
 [59.905431855598714 nm, 18.420162158120924 nm, 246.09423563049683 nm]
 [24.572857402685223 nm, 17.079711089908244 nm, 27.137908845124624 nm]
 [22.222703192699257 nm, 177.77658095653265 nm, 134.81396187256172 nm]

 julia> CellListMap.strip_value(x::Unitful.Quantity) = Unitful.ustrip(x) 

 julia> cl = CellList(x,box) # now x can be passed to CellList
CellList{3, Float64}
  5 real particles.
  5 cells with real particles.
  7 particles in computing box, including images.

CellListMap.translation_image — Method

translation_image(x::SVector{N,T},unit_cell_matrix,indices) where {N,T}

Translate vector x according to the unit_cell_matrix lattice vectors and the indices provided.

CellListMap.view_celllist_particles — Method

view_celllist_particles(cl::CellList)

Auxiliary function to view the particles of a computing box, including images created for computing purposes.

Example

julia> box = Box([ 100 50; 50 100 ],10);

julia> x = [ box.unit_cell_max .* rand(SVector{2,Float64}) for i in 1:1000 ];

julia> cl = CellList(x,box);

julia> p = CellListMap.view_celllist_particles(cl);

julia> using Plots

julia> scatter(Tuple.(p),label=nothing,xlims=(-10,180),ylims=(-10,180))

CellListMap.wrap_cell_fraction — Method

wrap_cell_fraction(x,unit_cell_matrix)

Obtaint the coordinates of x as a fraction of unit cell vectors, first positive cell. x is a vector of dimension N and cell a matrix of dimension NxN

Example

julia> unit_cell_matrix = [ 10 0
                            0 10 ];

julia> x = [ 15, 13 ];

julia> wrap_cell_fraction(x,unit_cell_matrix)
2-element Vector{Float64}:
 0.5
 0.3

CellListMap.wrap_relative_to — Method

wrap_relative_to(x,xref,sides::AbstractVector)

Wraps the coordinates of point x such that it is the minimum image relative to xref, for an Orthorhombic cell of which only the side lengths are provided.

CellListMap.wrap_relative_to — Method

wrap_relative_to(x,xref,box::Box{OrthorhombicCell,N,T}) where {N,T}

Wraps the coordinates of point x such that it is the minimum image relative to xref, given an Orthorhombic cell. This is slightly cheaper than for general cells.

CellListMap.wrap_relative_to — Method

wrap_relative_to(x, xref, unit_cell_matrix::SMatrix{N,N,T}) where {N,T}

Wraps the coordinates of point x such that it is the minimum image relative to xref.

CellListMap.wrap_relative_to — Method

wrap_relative_to(x,xref,box::Box{UnitCellType,N,T}) where {UnitCellType,N,T}

Wraps the coordinates of point x such that it is the minimum image relative to xref, given a general Box structure.

CellListMap.wrap_to_first — Method

wrap_to_first(x,unit_cell_matrix)

Wraps the coordinates of point x such that the returning coordinates are in the first unit cell with all-positive coordinates.

Example

```julia-repl julia> unitcellmatrix = [ 10 0 0 10 ];

julia> x = [ 15, 13 ];

julia> wraptofirst(x,unitcellmatrix) 2-element Vector{Float64}: 5.0 3.0000000000000004 ```

CellListMap.wrap_to_first — Method

wrap_to_first(x,box::Box)

Wraps the coordinates of point x such that the returning coordinates are in the first unit cell with all-positive coordinates, given the Box structure.

CellListMap.wrap_to_first — Method

wrap_to_first(x,box::Box{OrthorhombicCell,N,T}) where {N,T}

Wraps the coordinates of point x such that the returning coordinates are in the first unit cell with all-positive coordinates, given an Orthorhombic cell. This is slightly cheaper than for general cells.